Fix: Objects not loaded after repeated put #11
Similar bugs founded in my project.

It seems that Local<Value>:New(args[0]) will be freed by v8 in a very
bad way (which is, the underlying storage is freed, while the ptr is
still in hold).

So I just looked at the commit history and fixed it by using the
copy-to-local first strategy. Currently this seems work in my place, and
should work other place. It is safer, slower, but it make 0.4.* branch work.